Possible Fix for #17 Invisible Brain Slime
This is based on the theory that Brain Slimes are dying on the client and not the server, so they are still in the world making sounds and jumping on your head, but invisible.

Based on how setSlimeSize() works, it seems to me slimes with size 0 end up with 0/0 health. This resets size-0 slimes with 2/2 health for lack of a more elegant solution.

Unfortunately it doesn't explain the inconsistent nature of this bug, which also makes it hard to test. I'm merely hoping it's related.